update nhanvien controller

+++ DTOs/NhanVien/NhanVienResponseDto.cs
@@ -12,6 +12,6 @@ namespace ManagementHotel.DTOs.NhanVien
         public string? DiaChi { get; set; }
         public DateTime? NgaySinh { get; set; }
         public string? ChucVu { get; set; }
-        public decimal? Luong { get; set; }
+        public int? Luong { get; set; }
     }
 }
